Detangle POSIX time from nativelib
