+++ nativelib/src/main/resources/scala-native/time_nano.cpp
